+++ src/features/collage/components/BulkUploadSection.js
