+++ src/features/collage/components/CanvasCollagePreview.js
