+++ src/features/collage/components/CaptionEditor.js
