+++ src/features/collage/components/CollageLayoutComponents.js
