+++ src/features/collage/components/CollagePreview.js
