+++ src/features/collage/components/CollageResultDialog.js
