+++ src/features/collage/components/DisclosureCard.examples.md
