+++ src/features/collage/components/DisclosureCard.js
