+++ src/features/collage/components/EarlyAccessFeedback.js
@@ -23,30 +23,7 @@ import SendIcon from '@mui/icons-material/Send';
 import { useNavigate } from 'react-router-dom';
 import { UserContext } from '../../../UserContext';
 import { SnackbarContext } from '../../../SnackbarContext';
-
-// Utility function to hash username for localStorage
-const hashString = (str) => {
-  let hash = 0;
-  if (str.length === 0) return hash;
-  for (let i = 0; i < str.length; i += 1) {
-    const char = str.charCodeAt(i);
-    hash = ((hash * 33) - hash) + char;
-    hash = Math.imul(hash, 1); // Convert to 32bit integer
-  }
-  return Math.abs(hash).toString();
-};
-
-// Utility functions for localStorage preference management
-const getCollagePreferenceKey = (user) => {
-  if (!user?.userDetails?.email) return 'memeSRC-collage-preference-anonymous';
-  const hashedUsername = hashString(user.userDetails.email);
-  return `memeSRC-collage-preference-${hashedUsername}`;
-};
-
-const setCollagePreference = (user, preference) => {
-  const key = getCollagePreferenceKey(user);
-  localStorage.setItem(key, preference);
-};
+import { setCollagePreference } from '../utils/preferences';
 
 export default function EarlyAccessFeedback() {
   const theme = useTheme();
