+++ src/features/collage/config/CollageConfig.js
