+++ src/features/collage/config/layouts/FivePanelLayouts.js
