+++ src/features/collage/config/layouts/FourPanelLayouts.js
