+++ src/features/collage/config/layouts/LayoutUtils.js
