+++ src/features/collage/config/layouts/ThreePanelLayouts.js
