+++ src/features/collage/config/layouts/TwoPanelLayouts.js
