+++ src/features/collage/config/layouts/index.js
