+++ src/features/collage/steps/CollageImagesStep.js
