+++ src/features/collage/steps/CollageSettingsStep.js
