+++ src/features/collage/styled/CollageStyled.js
